backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9        # for Celery result backend
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Pin bcrypt: 4.1+ breaks passlib 1.7.4 backend init (login / verify_password).
bcrypt==4.0.1
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4                  # Valkey-compatible; redis-py client works with Valkey
openpyxl==3.1.3
weasyprint==61.2
# WeasyPrint 61.x is incompatible with pydyf 0.11+ (PDF.__init__ signature); see Kozea/WeasyPrint#2202.
pydyf>=0.5.0,<0.11
python-dateutil==2.9.0
httpx==0.27.0                 # For Superset guest token API calls
